feat: upgraded to using snapdeck library

+++ index.js
@@ -2,6 +2,7 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 const { Client, GatewayIntentBits, Partials } = require('discord.js');
+const snapdeck = require('snapdeck');
 const parseDeck = require('./parseDeck');
 
 const TOKEN = process.env.NEO_BOT_TOKEN; // Replace with your bot token
@@ -21,13 +22,22 @@ client.on('clientReady', () => {
 client.on('messageCreate', async (message) => {
 	if (message.author.bot) return;
 
-	if (/^# \(\d+/.test(message.content)) {
-		try {
-			await message.delete();
-			await message.channel.send(`Here's the deck <@${message.author.id}> shared:\n\`\`\`${parseDeck(message.content)}\`\`\``);
-		} catch (err) {
-			console.error('Failed to delete or send message:', err);
-		}
+	try {
+		const deckcode = snapdeck.extractDeckcode(message.content);
+		console.log(`extracted deckcode: ${JSON.stringify(deckcode)}`);
+		if (!deckcode) return;
+		const deck = await snapdeck.parseDeckcode(deckcode);
+		console.log(`parsed deck: ${JSON.stringify(deck)}`);
+		if (!deck) return;
+		const displayDeck = await snapdeck.generateDisplayString(deck.cards);
+		console.log(`generated display deck: ${displayDeck}`);
+		if (!displayDeck) return;
+
+		await message.delete();
+
+		await message.channel.send(`Here's the deck <@${message.author.id}> shared:\n\`\`\`${displayDeck}\`\`\``);
+	} catch (err) {
+		console.log(`error processing: ${message.content}\n${err}`);
 	}
 });
